Enable pyrenode3 on m-series
Initial commit to enable support for running pyrenode3 on m-series macs

The crux of the problem comes from some bad defaults in the
configuration of various support packages.

- By default the assumed runtime for pythonnet is 'mono'
- The system package path defaults when loading 'mono' from environment
contains only libraries for x86, i386
- dlopen is unable to load these kinds of libraries

all of this results in a pretty nasty error message upon even attempting
to import the pyrenode module

building the mono runtime from source overcomes the limitation of not
having dynamic libraries with the right architecture

then what's needed is to provide an option to explicitly load the
mono runtime on a specific path - rather than just by identifier

once the pythonnet runtime has been initialized, the subsequent calls to
load the Renode assemblies into the runtime will succeed*, allowing the
module to be imported.

one caveat to this being successful is the missing depedencies on
gtk-sharp and glib-sharp, which are not included in the source
installation of mono.  Rather than get those dependencies, simply
allow disabling the loading of the  Xwt assembly which introduces
them by environment variable.  Means certain graphics features won't be
enabled - which should be fine for running via a script

Need to come up with ways of making this more platform independent